feat: 整理 common
