+++ app/api/common.ts
@@ -1,5 +1,6 @@
 import { NextRequest } from "next/server";
 
+const API_DOMAIN = "http://192.168.31.73:8000";
 const OPENAI_URL = "api.openai.com";
 const DEFAULT_PROTOCOL = "https";
 const PROTOCOL = process.env.PROTOCOL ?? DEFAULT_PROTOCOL;
@@ -18,5 +19,29 @@ export async function requestOpenai(req: NextRequest) {
     },
     method: req.method,
     body: req.body,
+  });
+}
+
+export function commonFetch(url: string, options: any) {
+  return new Promise<any>((resolve, reject) => {
+    return fetch(`${API_DOMAIN}/api/${url}`, options).then((res: any) => {
+      console.log("handleFetch", res);
+
+      if (res.status === 204) {
+        return res.text().then((result: any) => {
+          resolve({ result, status: res.status });
+        });
+      }
+
+      if (res.status >= 200 && res.status <= 300) {
+        res.json().then((result: any) => {
+          resolve({ result, status: res.status });
+        });
+      } else {
+        res.json().then((result: any) => {
+          reject({ result, status: res.status });
+        });
+      }
+    });
   });
 }
